backend/requirements.txt
APScheduler==3.10.4
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==41.0.7
dnspython==2.7.0
email-validator==2.1.0
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
soupsieve==2.7
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
